backend/requirements.txt
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.20.0
pymongo==4.5.0
pydantic-settings==2.2.1
python-dotenv==0.21.1
SQLAlchemy==2.0.13
loguru==0.7.2
python-docx==1.1.0
openai==1.11.1
requests==2.31.0
email-validator==2.1.1
passlib==1.7.4
python-jose[cryptography]
boto3==1.29.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langchain-openai==0.0.5
langsmith==0.1.49
fastembed==0.2.6
pdfplumber==0.11.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
openai==1.11.1